test: wait for redirect to happen before parsing auth code
Our integration tests would sometimes fail if Connect took longer to
redirect from the grant screen to the redirect URI because our code did
not wait for the navigation to be complete. This would result in our
code attempting to parse out a code from the grant page URL which in
turn lead to a NPE.

The fix is to simply wait for the URL to be redirect URI before moving
on.