feat(fuel)!: Fuel data types from primitives to objects
This is the change to update the data types is fuel endpoint to allow nulls when one or two of the fuel signals is empty.

This change might break the customers when one of the signals return null from the Smartcar API server. Customer need to be aware to map correctly the null cases.